+++ src/lib/components/modal/Body.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+	import { twMerge } from 'tailwind-merge';
+	import { get_current_component } from 'svelte/internal';
+	import { forwardEventsBuilder, useActions, type ActionArray } from '../../actions';
+	export let use: ActionArray = [];
+	import { exclude } from '../../utils/exclude';
+	const forwardEvents = forwardEventsBuilder(get_current_component());
+
+	const defaultClass =
+		'first:rounded-t-md last:rounded-b-md px-4 py-5 sm:px-6 flex-grow overflow-y-auto';
+	$: finalClass = twMerge(defaultClass, $$props.class);
+</script>
+
+<div
+	class={finalClass}
+	use:useActions={use}
+	use:forwardEvents
+	{...exclude($$props, ['use', 'class'])}
+>
+	<slot />
+</div>