+++ src/lib/types/index.ts
@@ -3,6 +3,15 @@ import type { Action as LightboxAction } from './lightbox-action';
 import type { TableColumn } from './table-column';
 import type { TimelineItem } from './timeline-item';
 import type TwSizes from './twSizes';
+import type { SelectOption } from './select-option';
 import type { DropResult } from './drop-result';
 
-export { CarouselSlide, DropResult, LightboxAction, TableColumn, TimelineItem, TwSizes };
+export {
+	CarouselSlide,
+	DropResult,
+	LightboxAction,
+	SelectOption,
+	TableColumn,
+	TimelineItem,
+	TwSizes
+};