+++ src/routes/badge/+page.svelte
@@ -17,6 +17,7 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Default</Card.Header>
 		<Card.Content slot="content" class="p-4">
+			<Badge>Default</Badge>
 			<Badge type="info">Default</Badge>
 			<Badge type="success">Default</Badge>
 			<Badge type="warn">Default</Badge>
@@ -34,10 +35,20 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Custom</Card.Header>
 		<Card.Content slot="content" class="p-4">
-			<Badge class="bg-green-800 text-green-200 bg-opacity-100 dark:bg-opacity-100">Custom</Badge>
-			<Badge class="bg-blue-800 text-blue-200 bg-opacity-100 dark:bg-opacity-100">Custom</Badge>
-			<Badge class="bg-red-800 text-red-200 bg-opacity-100 dark:bg-opacity-100">Custom</Badge>
-			<Badge class="bg-amber-800 text-amber-200 bg-opacity-100 dark:bg-opacity-100">Custom</Badge>
+			<Badge
+				class="bg-green-800 dark:bg-green-800 text-green-200 bg-opacity-100 dark:bg-opacity-100"
+				>Custom</Badge
+			>
+			<Badge class="bg-blue-800 dark:bg-blue-800 text-blue-200 bg-opacity-100 dark:bg-opacity-100"
+				>Custom</Badge
+			>
+			<Badge class="bg-red-800 dark:bg-red-800 text-red-200 bg-opacity-100 dark:bg-opacity-100"
+				>Custom</Badge
+			>
+			<Badge
+				class="bg-amber-800 dark:bg-amber-800 text-amber-200 bg-opacity-100 dark:bg-opacity-100"
+				>Custom</Badge
+			>
 
 			<br />
 			<br />
@@ -51,6 +62,7 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Large</Card.Header>
 		<Card.Content slot="content" class="p-4">
+			<Badge size="large">Large</Badge>
 			<Badge type="info" size="large">Large</Badge>
 			<Badge type="success" size="large">Large</Badge>
 			<Badge type="warn" size="large">Large</Badge>
@@ -68,6 +80,10 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Indicator</Card.Header>
 		<Card.Content slot="content" class="p-4">
+			<Badge>
+				<Badge.Indicator slot="indicator" />
+				Indicator
+			</Badge>
 			<Badge type="info">
 				<Badge.Indicator slot="indicator" />
 				Indicator
@@ -97,11 +113,13 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Shape</Card.Header>
 		<Card.Content slot="content" class="p-4">
+			<Badge shape="rounded">Shape</Badge>
 			<Badge type="info" shape="rounded">Shape</Badge>
 			<Badge type="success" shape="rounded">Shape</Badge>
 			<Badge type="warn" shape="rounded">Shape</Badge>
 			<Badge type="error" shape="rounded">Shape</Badge>
 			<br />
+			<Badge shape="square">Shape</Badge>
 			<Badge type="info" shape="square">Shape</Badge>
 			<Badge type="success" shape="square">Shape</Badge>
 			<Badge type="warn" shape="square">Shape</Badge>
@@ -119,6 +137,10 @@
 	<Card bordered={false}>
 		<Card.Header slot="header">Close</Card.Header>
 		<Card.Content slot="content" class="p-4">
+			<Badge>
+				Close
+				<Badge.Close slot="close" on:click={() => console.log('clicked badge close')} />
+			</Badge>
 			<Badge type="info">
 				Close
 				<Badge.Close slot="close" on:click={() => console.log('clicked badge close')} />