+++ src/routes/checkbox/examples.ts
@@ -42,6 +42,12 @@ export const checkboxProps: Prop[] = [
 		prop: 'value',
 		type: 'string',
 		default: ''
+	},
+	{
+		id: '3',
+		prop: 'indeterminate',
+		type: 'boolean',
+		default: 'false'
 	}
 ];
 
@@ -94,7 +100,7 @@ export const example = `
    <CheckboxGroup.Checkbox name="cb-3" value="cb-3">
       <CheckboxGroup.Checkbox.Label slot="label">Checkbox-3</CheckboxGroup.Checkbox.Label>
    </CheckboxGroup.Checkbox>
-   <CheckboxGroup.Checkbox name="cb-4" value="cb-4">
+   <CheckboxGroup.Checkbox indeterminate name="cb-4" value="cb-4">
       <CheckboxGroup.Checkbox.Label slot="label">Checkbox-4</CheckboxGroup.Checkbox.Label>
    </CheckboxGroup.Checkbox>
 </CheckboxGroup>`;