+++ src/routes/types/examples.ts
@@ -121,3 +121,10 @@ type TwSizes =
 	| '80'
 	| '96';
 `;
+
+export const selectOption = `
+interface SelectOption {
+	[key: string]: any;
+	value: string;
+}
+`;